semi-final-commit

+++ src/main/java/com/web/project/controller/LoaiSanPhamController.java
@@ -3,6 +3,7 @@ package com.web.project.controller;
 import java.io.IOException;
 import java.util.List;
 
+import com.web.project.service.SanPhamService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.repository.query.Param;
@@ -26,6 +27,9 @@ import com.web.project.service.NhanVienService;
 public class LoaiSanPhamController {
 	@Autowired
 	private LoaiSanPhamService service;
+
+	@Autowired
+	private SanPhamService sanPhamService;
 	
 	@GetMapping("/loaisanpham")
 	public String listAll(Model model) {
@@ -61,19 +65,29 @@ public class LoaiSanPhamController {
 	@GetMapping("/loaisanpham/new")
 	public String form(Model model) {
 		model.addAttribute("loai", new LoaiSanPham());
-		model.addAttribute("pageTitle", "Them loai san pham");
+		model.addAttribute("pageTitle", "Thêm loại sản phẩm");
 		return "loaisanpham/loaisanpham_form";
 	}
 	
 	@GetMapping("/loaisanpham/{id}/enabled/{enabled}")
 	public String updateStatus(@PathVariable("id")Integer id , @PathVariable("enabled")boolean trangThai ,RedirectAttributes re) {
 		service.updateTrangThai(id, trangThai);
-		re.addFlashAttribute("message", "Cap nhat trang thai thanh cong");
+		re.addFlashAttribute("message", "Cập nhật trạng thái thành công !!");
 		return "redirect:/loaisanpham";
 	}
 	
 	@PostMapping("/loaisanpham/save")
-	public String save(LoaiSanPham loaisanpham , RedirectAttributes re , @RequestParam("image") MultipartFile multipartFile) throws IOException {
+	public String save(LoaiSanPham loaisanpham ,
+					   Model model ,
+					   @RequestParam("image") MultipartFile multipartFile ,
+					   RedirectAttributes re)
+			throws IOException {
+		if(!service.checkTenUnique(loaisanpham.getId(),loaisanpham.getTen())){
+			model.addAttribute("message" , "Tên loại sản phẩm này đã được đặt");
+			model.addAttribute("loai" ,loaisanpham);
+			model.addAttribute("pageTitle", loaisanpham.getId() != null ? "Sửa loại sản phẩm" : "Thêm lọai sản phẩm");
+			return "loaisanpham/loaisanpham_form";
+		}
 		if(!multipartFile.isEmpty()) {
 			String fileName = StringUtils.cleanPath(multipartFile.getOriginalFilename());
 			loaisanpham.setPhoto(fileName);
@@ -85,7 +99,7 @@ public class LoaiSanPhamController {
 			if(loaisanpham.getPhoto().isEmpty()) loaisanpham.setPhoto(null);
 			service.save(loaisanpham);
 		}
-		re.addFlashAttribute("message", "Loai San pham da duoc luu thanh cong");
+		re.addFlashAttribute("message", "Loại sản phẩm nầy đã được lưu thành công !!");
 		return "redirect:/loaisanpham";
 	}
 	@GetMapping("/loaisanpham/edit/{id}")
@@ -94,7 +108,7 @@ public class LoaiSanPhamController {
 		try {
 			LoaiSanPham loai = service.getById(id);
 			model.addAttribute("loai", loai);
-			model.addAttribute("pageTitle", "Sua loai san pham");
+			model.addAttribute("pageTitle", "Sửa loại sản phẩm");
 			return "loaisanpham/loaisanpham_form";
 		} catch (LoaiSanPhamNotFoundException e) {
 			re.addFlashAttribute("message", e.getMessage());
@@ -106,6 +120,10 @@ public class LoaiSanPhamController {
 	@GetMapping("/loaisanpham/delete/{id}")
 	public String delete (@PathVariable("id") Integer id , RedirectAttributes re) {
 		try {
+			if(!sanPhamService.canDeleteLoai(id)){
+				re.addFlashAttribute("message" , "Không thể xóa loại sản phẩm có id là:" + id);
+				return "redirect:/loaisanpham";
+			}
 			service.delete(id);
 		} catch (LoaiSanPhamNotFoundException e) {
 			re.addFlashAttribute("message", e.getMessage());
